mqtt_app/requirements.txt
paho-mqtt
mysql-connector
